lib/xos-config/requirements.txt
PyYAML~=3.12
pykwalify~=1.6.0
